backend/Django/requirements.txt
asgiref==3.3.1
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-default-imagefield==0.5.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
humanize==3.3.0
install==1.3.4
pgcli==1.9.0
pgspecial==1.11.5
Pillow==8.1.2
prompt-toolkit==1.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.8.1
PyJWT==2.0.1
pytz==2021.1
setproctitle==1.2.2
six==1.15.0
sqlparse==0.2.4
tabulate==0.8.9
terminaltables==3.1.0
wcwidth==0.2.5
whitenoise==5.2.0
